+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Features
+* `exec` supports a callback argument. (#290)
+
 ### Fixes
 * Don't modify objects passed to or returned from the internal datastore. (#281 @jeffkenney)
 * Fix hash get commands for missing hashes. (#284 @jeffkenney)